requirements.txt
# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Model Interpretability
shap>=0.40.0

# Jupyter and Interactive Computing
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Progress Bars and Utilities
tqdm>=4.62.0

# Optional: Geospatial (if working with GEE data locally)
# earthengine-api>=0.1.300
# geopandas>=0.10.0
# Deep Learning
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0
